% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acct.R
\name{AcctUI}
\alias{AcctUI}
\title{按纽生成生成器界面}
\usage{
AcctUI(
  tabTitle = "科目",
  colTitles = c("操作区", "", "显示区"),
  widthRates = c(6, 6, 12),
  func_left = AcctUI_left,
  func_right = AcctUI_right,
  func_bottom = AcctUI_bottom
)
}
\arguments{
\item{tabTitle}{标题}

\item{colTitles}{主页标题}

\item{widthRates}{左右比例}

\item{func_left}{左函数}

\item{func_right}{右函数}

\item{func_bottom}{下面一栏}
}
\value{
返回值
}
\description{
按纽生成生成器界面
}
\examples{
AcctUI()
}
